requirements.txt
fastapi
uvicorn
aiosqlite
sqlmodel
sqlalchemy
pytest
pytest-asyncio
python-decouple
requests
